requirements.txt
Flask==3.1.0
PyYAML==6.0.2
pytest==8.3.4
